splitting ProductList.jsx
